+++ packages/typir/src/graph/graph-algorithms.ts
@@ -5,7 +5,8 @@
  ******************************************************************************/
 
 import { TypirServices, TypirSpecifics } from '../typir.js';
-import { TypeEdge } from './type-edge.js';
+import { assertUnreachable } from '../utils/utils.js';
+import { RelationInformation, TypeEdge } from './type-edge.js';
 import { TypeGraph } from './type-graph.js';
 import { Type } from './type-node.js';
 
@@ -14,9 +15,9 @@ import { Type } from './type-node.js';
  * All algorithms are robust regarding cycles.
  */
 export interface GraphAlgorithms {
-    collectReachableTypes(from: Type, $relations: Array<TypeEdge['$relation']>, filterEdges?: (edgr: TypeEdge) => boolean): Set<Type>;
-    existsEdgePath(from: Type, to: Type, $relations: Array<TypeEdge['$relation']>, filterEdges?: (edgr: TypeEdge) => boolean): boolean;
-    getEdgePath(from: Type, to: Type, $relations: Array<TypeEdge['$relation']>, filterEdges?: (edgr: TypeEdge) => boolean): TypeEdge[];
+    collectReachableTypes(from: Type, $relations: RelationInformation[], filterEdges?: (edgr: TypeEdge) => boolean): Set<Type>;
+    existsEdgePath(from: Type, to: Type, $relations: RelationInformation[], filterEdges?: (edgr: TypeEdge) => boolean): boolean;
+    getEdgePath(from: Type, to: Type, $relations: RelationInformation[], filterEdges?: (edgr: TypeEdge) => boolean): TypeEdge[];
 }
 
 export class DefaultGraphAlgorithms<Specifics extends TypirSpecifics> implements GraphAlgorithms {
@@ -26,20 +27,20 @@ export class DefaultGraphAlgorithms<Specifics extends TypirSpecifics> implements
         this.graph = services.infrastructure.Graph;
     }
 
-    collectReachableTypes(from: Type, $relations: Array<TypeEdge['$relation']>, filterEdges?: (edgr: TypeEdge) => boolean): Set<Type> {
+    collectReachableTypes(from: Type, $relations: RelationInformation[], filterEdges?: (edgr: TypeEdge) => boolean): Set<Type> {
         const result: Set<Type> = new Set();
         const remainingToCheck: Type[] = [from];
 
         while (remainingToCheck.length > 0) {
             const current = remainingToCheck.pop()!;
-            const outgoingEdges = $relations.flatMap(r => current.getOutgoingEdges(r));
-            for (const edge of outgoingEdges) {
-                if (edge.cachingInformation === 'LINK_EXISTS' && (filterEdges === undefined || filterEdges(edge))) {
-                    if (result.has(edge.to)) {
+
+            for (const $relation of $relations) { // check the $relations in the given order
+                for (const { otherEnd } of this.calculateRelevantEdges($relation, current, filterEdges)) {
+                    if (result.has(otherEnd)) {
                         // already checked
                     } else {
-                        result.add(edge.to); // this type is reachable
-                        remainingToCheck.push(edge.to); // check it for recursive conversions
+                        result.add(otherEnd); // this type is reachable
+                        remainingToCheck.push(otherEnd); // check it for recursive conversions
                     }
                 }
             }
@@ -48,7 +49,7 @@ export class DefaultGraphAlgorithms<Specifics extends TypirSpecifics> implements
         return result;
     }
 
-    existsEdgePath(from: Type, to: Type, $relations: Array<TypeEdge['$relation']>, filterEdges?: (edgr: TypeEdge) => boolean): boolean {
+    existsEdgePath(from: Type, to: Type, $relations: RelationInformation[], filterEdges?: (edgr: TypeEdge) => boolean): boolean {
         const visited: Set<Type> = new Set();
         const stack: Type[] = [from];
 
@@ -56,10 +57,9 @@ export class DefaultGraphAlgorithms<Specifics extends TypirSpecifics> implements
             const current = stack.pop()!;
             visited.add(current);
 
-            const outgoingEdges = $relations.flatMap(r => current.getOutgoingEdges(r));
-            for (const edge of outgoingEdges) {
-                if (edge.cachingInformation === 'LINK_EXISTS' && (filterEdges === undefined || filterEdges(edge))) {
-                    if (edge.to === to) {
+            for (const $relation of $relations) { // check the $relations in the given order
+                for (const { otherEnd } of this.calculateRelevantEdges($relation, current, filterEdges)) {
+                    if (otherEnd === to) {
                         /* It was possible to reach our goal type using this path.
                          * Base case that also catches the case in which start and end are the same
                          * (is there a cycle?). Therefore it is allowed to have been "visited".
@@ -67,11 +67,11 @@ export class DefaultGraphAlgorithms<Specifics extends TypirSpecifics> implements
                          */
                         return true;
                     }
-                    if (!visited.has(edge.to)) {
+                    if (!visited.has(otherEnd)) {
                         /* The target node of this edge has not been visited before and is also not our goal node
                          * Add it to the stack and investigate this path later.
                          */
-                        stack.push(edge.to);
+                        stack.push(otherEnd);
                     }
                 }
             }
@@ -81,18 +81,17 @@ export class DefaultGraphAlgorithms<Specifics extends TypirSpecifics> implements
         return false;
     }
 
-    getEdgePath(from: Type, to: Type, $relations: Array<TypeEdge['$relation']>, filterEdges?: (edgr: TypeEdge) => boolean): TypeEdge[] {
+    getEdgePath(from: Type, to: Type, $relations: RelationInformation[], filterEdges?: (edgr: TypeEdge) => boolean): TypeEdge[] {
         const visited: Map<Type, TypeEdge|undefined> = new Map(); // the edge from the parent to the current node
         visited.set(from, undefined);
-        const stack: Type[] = [from];
+        const stack: Type[] = [from]; // stores the next types to investigate
 
         while (stack.length > 0) {
             const current = stack.pop()!;
 
-            const outgoingEdges = $relations.flatMap(r => current.getOutgoingEdges(r));
-            for (const edge of outgoingEdges) {
-                if (edge.cachingInformation === 'LINK_EXISTS' && (filterEdges === undefined || filterEdges(edge))) {
-                    if (edge.to === to) {
+            for (const $relation of $relations) { // check the $relations in the given order
+                for (const { edge, myEnd, otherEnd } of this.calculateRelevantEdges($relation, current, filterEdges)) {
+                    if (otherEnd === to) {
                         /* It was possible to reach our goal type using this path.
                          * Base case that also catches the case in which start and end are the same
                          * (is there a cycle?). Therefore it is allowed to have been "visited".
@@ -100,20 +99,20 @@ export class DefaultGraphAlgorithms<Specifics extends TypirSpecifics> implements
                          */
                         const result: TypeEdge[] = [edge];
                         // collect the path of used edges, from "to" back to "from"
-                        let backNode = edge.from;
+                        let backNode = myEnd;
                         while (backNode !== from) {
                             const backEdge = visited.get(backNode)!;
                             result.unshift(backEdge);
-                            backNode = backEdge.from;
+                            backNode = backEdge.to === backNode ? backEdge.from : backEdge.to; // handle bidirectional edges defined for the "wrong" direction
                         }
                         return result;
                     }
-                    if (!visited.has(edge.to)) {
+                    if (!visited.has(otherEnd)) {
                         /* The target node of this edge has not been visited before and is also not our goal node
                          * Add it to the stack and investigate this path later.
                          */
-                        stack.push(edge.to);
-                        visited.set(edge.to, edge);
+                        stack.push(otherEnd);
+                        visited.set(otherEnd, edge);
                     }
                 }
             }
@@ -123,4 +122,23 @@ export class DefaultGraphAlgorithms<Specifics extends TypirSpecifics> implements
         return [];
     }
 
+    protected calculateRelevantEdges($relation: RelationInformation, current: Type, filterEdges?: (edgr: TypeEdge) => boolean): Array<{ edge: TypeEdge, myEnd: Type, otherEnd: Type }> {
+        if ($relation.direction === 'Unidirectional') {
+            return current.getOutgoingEdges($relation.$relation)
+                .filter(e => e.cachingInformation === 'LINK_EXISTS' && (filterEdges === undefined || filterEdges(e)))
+                .map(e => ({ edge: e, myEnd: e.from, otherEnd: e.to }));
+        }
+        if ($relation.direction === 'Bidirectional') {
+            // for bidirectional edges, both outgoing and incoming edges need to be checked, while to and from are swapped
+            return [
+                ...current.getOutgoingEdges($relation.$relation)
+                    .filter(e => e.cachingInformation === 'LINK_EXISTS' && (filterEdges === undefined || filterEdges(e)))
+                    .map(e => ({ edge: e, myEnd: e.from, otherEnd: e.to })),
+                ...current.getIncomingEdges($relation.$relation)
+                    .filter(e => e.cachingInformation === 'LINK_EXISTS' && (filterEdges === undefined || filterEdges(e)))
+                    .map(e => ({ edge: e, myEnd: e.to, otherEnd: e.from })),
+            ];
+        }
+        assertUnreachable($relation.direction);
+    }
 }