+++ packages/typir/src/services/subtype.ts
@@ -92,7 +92,7 @@ export class DefaultSubType<Specifics extends TypirSpecifics> implements SubType
 
     getSubTypeResult(subType: Type, superType: Type): SubTypeResult {
         // search for a transitive sub-type relationship
-        const path = this.algorithms.getEdgePath(subType, superType, [SubTypeEdge]);
+        const path = this.algorithms.getEdgePath(subType, superType, [{ $relation: SubTypeEdge, direction: 'Bidirectional' }]);
         if (path.length >= 1) {
             return <SubTypeSuccess>{
                 $result: SubTypeResult,
@@ -144,7 +144,7 @@ export class DefaultSubType<Specifics extends TypirSpecifics> implements SubType
 
         // check for cycles
         if (actualOptions.checkForCycles) {
-            const hasIntroducedCycle = this.algorithms.existsEdgePath(subType, subType, [SubTypeEdge]);
+            const hasIntroducedCycle = this.algorithms.existsEdgePath(subType, subType, [{ $relation: SubTypeEdge, direction: 'Unidirectional' }]);
             if (hasIntroducedCycle) {
                 throw new Error(`Adding the sub-type relationship from ${subType.getIdentifier()} to ${superType.getIdentifier()} has introduced a cycle in the type graph.`);
             }